Added files to SetupFiles directory for those who use bash as their login shell.  Made changes to make_SET_ME_UP to support this.  After running make_SET_ME_UP the user needs to source the appropriate file in SetupFiles: SET_ME_UP.csh or SET_ME_UP.bash.
Altered files in the coda subdirectory to make the evio library 64-bit safe.
Rather than casting pointers-to-structs to ints to pass between functions it
is better to just pass the struct pointers.  (Pointers on 64-bit systems are 8
bytes long whereas ints are 4 bytes long.  Casting a pointer to an int results
in the loss of significant information.)

Commented out some preprocesser directives in evio.C that defined certain
functions if VXWRKS was not defined.  Does not appear to affect compilation on
64-bit Fedora 8 or on the JLab CUE.

Changed the Makefile in coda subdirectory so it compiles evio.o and
swap_utils.o like it compiles all the other object files.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@76 b9700a98-9937-0410-b777-f318d6eb5bff