This update consists of  changes to Qweak tracking code on QwHitContainer class, QwDriftChamber class and it's sub classes and QwAnalysis.cc
This code generates a grand hit list(QwHitContainer class object with hits from region 2 and region 3) from region 2 and region 3 hits information. Since region 3 is under construction, we have generated a set of fake hits for region 3, which are nothing but same hits from region 2. QwHitContainer hit list is sorted and the class has the ability to retrieve set of hits for given criteria.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@86 b9700a98-9937-0410-b777-f318d6eb5bff