The module responsible for reading in the test events (Qevent) intermittently caused segmentation faults.  After minor modifications it reliably reads in the event files.  Two new event files, with events in region 2 and with events in region 3, were added.
The geometry file was updated to include region 2 detectors, and minor changes
to the region 3 detectors were applied (still from Burnham's tarball).

Due to an uninitialized variable (in treesearch) the previous version of the
tracking code crashed while doing the tree search.  It now works on the
event file 1000.r3.events (first link this file to qweak.event).

A small change in the makefile because some platforms require a run-time
library path (CXXFLAGS += -Wl,-rpath).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@110 b9700a98-9937-0410-b777-f318d6eb5bff