Sorry for the huge monolithic commit.  I wish I could have split it up better.
- Removal of enum.h (and Burnham's old enumerators): Replaced where necessary with enumerators in QwTypes.h.  Please do a make config 
to ensure that the proper dependencies are in place (or it will try to find enum.h).  This is the cause for changes in many of the 
Tracking code files.

- Renamed kUP and kDOWN to kPackageUp and kPackageDown.  Similar name changes for EQwDetectorType.  Occurrences in other code also 
changed.  Renamed kNumPlanes to kNumTypes (at every instance of rcTreeRegion, rcDETRegion).  Again this caused many small changes 
e.g. in the Parity main().

- Addition of QwHit tree pattern search functions: Hit patterns are now generated from the QwHitContainer list (but still far from 
optimal).  Legacy Hit class is still used in various places (e.g. in treelines where a list of hits is kept), but will be phased out.  
treedo::rcTreeDo takes now a QwHitContainer as argument instead of the event number.

- Change to qweak.options: generic drift chamber (d) is now either HDC or VDC.  Unless you use a local qweak.options file (which is 
very trakcing specific), you probably don't care about this.

As Rakitha pointed out, some events in region 3 still make it crash.  Actually, all events in region 3 make it crash... :-|



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@154 b9700a98-9937-0410-b777-f318d6eb5bff