Additional functionality for the QwAnalysis framework: a Qw-ROOT interface
With QwRoot and QwControl it is possible to have a ROOT prompt inside
a QwAnalysis session.  It is running in a different thread (as are the
analyzer and the dataserver, two components that I had to add) and ROOT
is taking care of all the nasty details like mutexes.

I added a simple example QwControlTest.cc that does something similar
to what QwTracking.cc does: loop over 1000 events and do the tracking
(or at least try).  It defines a ROOT histogram and puts the chi^2 of
the treelines in there (I chose that because there are some issues with
it, but anything would be possible).  After the tracking is done (or
earlier, if you are fast enough to type 'new TBrowser' before it ends)
you can look at what the chi^2 distribution is (in root -> ROOT memory).

The philosophy of the example is different in that it pushes all the
functionality into classes rather than in a main() function.  This was
necessary to make it multithreaded.

Getting this compiled is a bit difficult sometimes.  The dictionary
generation with cint is not very reliable (probably I have a bug in
a header file).  Also, the dependencies are not all found.  So you
will have the best results with 'EXES=qwcontroltest qwanalysis qwtracking',
although even then it sometimes fails to compile.  A distclean sometimes
fixes this.

Because QwControl starts a QwTrackingAnalyzer, it ties the three parts
of QwAnalysis again together (sorry Paul).  But this really only happens
in a single line, everything else is based on a virtual VQwAnalyzer class.
There must be a way to keep this separate: how do you instantiate an object
that inherits from a base class without specifying the class at compile
time?

Finally, this is nothing more than a 'proof of concept'.  I think it might
be a useful step in the direction of online monitoring software, but it
would need more stability.

New classes:
* QwControl: created in main(), could do command line parsing and option
  file (QwParameterFile) reading, starts other components, and then
  'becomes' the command prompt while waiting for input
* QwRoot: main object that holds the rest together, interfaces between
  the dataserver and analyzer
* QwTrackingDataserver (inherits from VQwDataserver): reads the ASCII file
  as a QwEventBuffer, produces QwHitContainers with Rakitha's routines
* QwTrackingAnalyzer (inherits from VQwAnalyzer): reads geometry, options,
  creates a QwTrackingWorker to do the tracking, and can fill histos
  afterwards



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@185 b9700a98-9937-0410-b777-f318d6eb5bff