This new class QwTreeEventBuffer should get us on the way to reading ROOT trees produced with the QweakSimG4 Monte Carlo package.  The present version should handle region 3 track (correct geometry info still missing, see below), but for some reason I could not get it to read the tree values from the ROOT file yet, so it is still a proof-of-concept. Distance from track to wire is calculated by closest approach here but can be made more realistic, taking into account nonlinear time-to-distance effects and varying resolution.
Region 2 is not implemented, but straightforward to add.  The problem is 
that only one front and one back wireplane in region 2 are activated in 
QweakSimG4.  I am not sure why, so unless there was a pertinent reason it
would be easy to change (it would make QweakSimEventAction even longer).

QwTreeEventBuffer inherits from QwEventBuffer and can be used just like
the QwASCIIEventBuffer:
e.g.
  QwTreeEventBuffer treebuffer(string filename);
  while (treebuffer.GetNextEvent() > 0) {
    // Get hit list
    QwHitContainer* hitlist = treebuffer.GetHitList();
  }
It does not fill the hits in rcTreeRegion, so until the tracking code is 
switched over to QwHits completely it is not very useful yet.  It also 
does not use rcDETRegion because I want to wait until we have a good 
std::vector<QwDetectorInfo> container through which we can access 
specific detectors.

The QweakSim.root file consists of 143 events with a 3-fold trigger (hit 
in HDC, VDC-front, VDC-back) out of 1000 simulated events.  You will 
need to modify QweakSimEventAction if you want to play with other types 
of events.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@202 b9700a98-9937-0410-b777-f318d6eb5bff