Thanks to Jie the reading in of Geant4 MC simulated events works.  I replaced the temporary code that determined the hit wires with something more realistic for region 3: the track position when entering and exiting each VDC chamber (active volume) is determined, all wires in between are set to be hit, with drift distance linearly changing between enter and exit point (this always results in one hit with zero drift time, but that is something we can live with for now).
This histlist can now be used as input for the tracking worker.  I modified/renamed 
the example code to QwSimTracking.cc, where it will still remain somewhat of a 
proof-of-concept: the tracking code still depends on the hit information being 
available in the old data structure, so those modifications will have to be performed 
soon.  But at least it can serve there also as a benchmark for that Hit -> QwHit 
transition (a goal!).

Region 2 is not yet implemented in this MC tree reader.  It still needs small changes 
to the QweakSimG4 MC to include all the planes.  Once that's done, it is trivial to add
region 2 to the tree reader, because only one wire is hit for every plane.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@238 b9700a98-9937-0410-b777-f318d6eb5bff