Rather large list of modified files, but only small modifications:
- Cleaned up some more memory leaks so the tracking can run over more than a few thousand events

- Added a dependency on boost::numeric::ublas which is standard in boost 1.33.1 and included on RHEL5, no action required if you are on a recent platform. uBLAS is a 
linear algebra package that is now used for the chi2 fit.  It is slightly slower for small matrices (2x2), but is much easier to debug.  Also, there is some 4x4 LU and 
Cholesky decomposition code that I want to get rid of in favor of the uBLAS functions.

- The vector objects in ublas were conflicting with the STL std::vector due to a prevalence of using namespace std lines all over the include files. I replaced them with 
(usually) using std::cout and using std::endl.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@298 b9700a98-9937-0410-b777-f318d6eb5bff