This commit allows for writing and reading of QwHitContainers to/from ROOT trees.
Although an intermediate TOrdCollection is used, I am fairly confident that this 
should work also on a regular QwHitContainer.  The main problem with STL containers 
seems to be with rootcint that cannot handle the complicated headers (more than one 
level of nested #ifdef #endif throws it off).  Most of those headers could be moved 
out of QwHitContainer.h into QwHitContainer.cc, leaving only attributes and 
signatures in the header file.  This should make rootcint happy.

QwHitContainerWrite.cc and QwHitContainerRead.cc respectively write and read a ROOT 
file called hitlist.root that contains a TOrdCollection named fQwHitContainer with 
all the hits. In ROOT you can browse using a TBrowser.  I am not a ROOT guru, so I 
couldn't figure out how to load the QwHit dictionary file (which, by the way, should 
now be svn:ignore'd).

Basically, the following is how to write the ROOT file:

    // Get hit list as QwHitContainer (new hitlist)
    QwHitContainer* hitlist = treebuffer->GetHitList(fEvtNum);

    // Replace rootlist with QwHitContainer
    QwHitRootContainer* rootlist->Replace(hitlist);

    // Save the event to tree
    tree->Branch("hits","QwHitRootContainer",&rootlist);
    tree->Fill();

And this is how to read it:

    TBranch* branch = tree->GetBranch("hits");
    QwHitRootContainer* rootlist = 0;
    branch->SetAddress(&rootlist);

    // Read next event
    branch->GetEntry(fEvtNum);

    // Read into the QwHitContainer format
    QwHitContainer* hitlist = rootlist->Read();

I agree that the name of the QwHitRootContainer object and its methods could have 
been chosen more sensibly.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@315 b9700a98-9937-0410-b777-f318d6eb5bff