Changes in VQWK to allow for an initial parity mock data implementation: - SetEventData allows to write the blocks to the VQWK ADC (and puts the   data structure in an internally consistent state) - GetEventBuffer retrieves the buffer that should go into the CODA file   (minus the headers of course, that will need to be added by the rest) This VQWK modification passes self-consistency tests: creating a buffer (in memory) results in the correct results on read-back.
In QwBeamLine I added a GetBCM and GetBPMStripline method (by name) so
specific elements can be addressed from the main routine.  This will be
necessary to write the correlated/random data to select elements.  This
will let you use the same map/pedestal files for creating and analyzing
parity mock data.

Small changes to QwBCM to pass through the artificial data to the under-
lying VQWK objects.  I plan to make this more physical, so that instead
of QwBCM::SetEventData(double VQWK_block[4]) you can pass something that
looks more like a current, and similarly a position for the QwBPMStrip-
lines.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@348 b9700a98-9937-0410-b777-f318d6eb5bff