Modification for the parity mock data generator.
VQwSubsystem defines now RandomizeEventData to put perfectly 
random data in the event structure.  Derived subsystems are 
supposed to make this more useful (add asymmetries and 
correlations).  EncodeEventData then returns the coda buffer for 
the subsystem.  RandomizeEventData is also added in the subsystem 
array and loops as expected over all the subsystems.

I added a couple of functions to include CODA prestart, go, pause 
and end events in the stream (they contain the run number etc, and 
I thought they were more important than they actually are).

QwVQWK_Channel now has an internal randomness generator (defined 
static, to avoid that all ADC channels get exactly the same stream 
of high quality random numbers...).  You can set the mean and 
width of the normal distribution using SetRandomEventParameters, 
and higher level subsystems (for example QwBCM and QwBPMStripline) 
are supposed to convert this to/from something useful (for example 
current and coordinates).

A unit conversion bug made the hardware and software sum different 
by one bit for the majority of generated events, and they were all 
thrown out by the parity analyzer.  Now they are really the same.

The example generator code now generates a CODA data file 
according to mock_qweak_beamline.map with uncorrelated random 
events.  It contains only BCMs and BPMs (but probably only the 
quartz bars need to be added).  The example analysis code does the 
analysis.

The next step is (still) adding the helicity data stream.  The 
helicity is not allowed to be random so from the second law of 
thermodynamics follows that this will be more difficult to 
implement.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@358 b9700a98-9937-0410-b777-f318d6eb5bff