I added the helicity to the parity mock data.  It isn't used yet in any asymmetry, but that would be a trivial addition, probably through RandomizeEventData(helicity) and SetRandomEventAsymmetry(double).
The helicity predictor (RunPredictor) had to be made public instead of protected,
because the cross check on the reported helicity (non-existent during mock data
generation) re-set the seed every time.  The seed can be specified with SetFirst24Bits().

Note: this code will only work with dolocalhelicity = kTRUE and dolocalhelicity2 = kFALSE,
and with the corresponding section in the qweak_helicity.map file.  Also, the uninitialized
variables which were forced to zero in a previous commit were again released.  This causes
annoying compilation warnings, but it is necessary for dolocalhelicity = kTRUE and I didn't
try to fix this behavior.  Please do not revert this warning without fixing the actual problem.

Segmentation faults due to a negative array index were 'cured' by forcing it to zero
in QwHelicityPattern.

Example:
>  qwmockdatagenerator -r 1 -e 0:1000
>  qwmockdataanalysis -r 1

It prints the helicity pattern while generating and again while analyzing (after 24 cycles),
so you can verify that it really is consistent.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@376 b9700a98-9937-0410-b777-f318d6eb5bff