Warning: following modifications change (private) interface of QwQuartzBar.
The QwQuartzBar was until now implemented slightly differently from the
other subsystems: it used a vector of *pointers* to VQWK modules, rather than
a vector of actual VQWK modules.  This causes issues with memory management
that can easily be avoided by using a vector of objects (and keeps the design
of this subsystem in line with others, e.g. QwBeamLine).  By modifying the
QwQuartzBar to usea vector of VQWK modules a number of memory leaks were removed.

Sorry if I cause some minor head aches for some people, but since there were no
commits to QwQuartzBar in the past 3 months, I assume that only few people are
affected.  Also, there could be good reasons for using an array of pointers
(though I couldn't see why they would apply in this case), so if there is some
functionality that would not be possible in this new scheme, feel free to revert.

Also implemented the Copy and Compare stub for use in the mock parity data
and parity histogram generation.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@387 b9700a98-9937-0410-b777-f318d6eb5bff