This might come as a surprise to some, but here is the Compton detector analysis code. Being on shift in Hall A, and having to see poor Compton experts suffer with a badly documented code is quite inspiring in a way...
This implements (partly) read-out of the SIS3320 sampling ADC (incidentally also used
in Hall A).  Right now, the support is still very limited (it is only going to read
part of a Compton CODA file), but the framework is put into place.  Originally, we
talked about taking the Hall A framework and using it to read our data (that worked
well at HIgS), but I think we have a more flexible solution here.

So, what is implemented?
- SIS3320 for the photon detector, V775 is already present and the V792 should be similar
- sums, differences, ratios,...
- framework for photon and electron detectors to fill in (loads in the QwSubsystemArrayParity)
- initial attempts at channel and pedestal map reading (*)
- initial support for mock data (that's the only way to test it, apart from HIgS data)
What needs implementing?
- electron detector, and thus V1495 read-out
- everything else ;-)

Other reasons why I think this will be easier than writing our own Compton analyzer:
- no need to reinvent the wheel/reimplement delayed helicity reporting
- inclusion in the Qweak analysis, that's the goal after all

Why this could be a bad idea:
- Qweak will end, the Compton is supposed to live on (in 2015 we will need to either
revive the Qweak code or fork the Compton code out of it).  That is, however, further
in the future and not so pressing as having a Compton analyzer ready by next year :-/

(*) Is there a way to get this shifted up in the class hierarchy?  Can we define a map
from string to functionality?  Can we support basic structure in the base classes and
allow for extensions in the children?  I feel like I am wasting my time writing code
that already exists almost the same in every other subsystem...


Note to Paul: Could you re-run doxygen?  I added some docs in the Compton code which
should make the structure clearer in the Compton group.  Thanks!



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@395 b9700a98-9937-0410-b777-f318d6eb5bff