Preliminary Event Cut Functionality for Beam line devices ========================================================
Note : The event cuts depends on the beam current and positions. Current values are set for data run 752 and 755


Beam line event cuts based on BCM and BPMStripline devices. The event cuts are set on a text file "qweak_beamline_eventcuts.in". The Applying Event cut can be turned off by setting bEventCut=kFALSE in the QwAnalysis_Beamline.cc

Before applying event cuts hardware checks were performed on the beam line devices. In the routine
VQwSubsystem::ProcessEvent() hardware checks were applied by using the IsGoodEvent() routine. Following hardware
checks were performed,

1. check for the sample size
2. ADC sequence number check
3. consistency between the HW_sum and the SW_sum
4. Check for ADCs returning the same value for consecutive events.
 and finally consistency within BPMStripline wires for above values.

The event cuts are then applied through VQwSubsystem::ApplySingleEventCuts(). Event cuts are fed in to the
analyzer currently by the text file. In the event cut file if the limits are set to zero then that device is ignored from event cut check. If the sample size is set to zero then that device is ignored from the hardware check.
Currently I'm checking event cuts only on qwk_bcm0l02 and qwk_qpd. The BPMStripline, qwk_1i06 is ignored from
all the check since it has generating an error consistently for all the events in the run. For the rest of the
devices only the hardware check is performed. 

For the devices that being checked for the event cuts <q>, <q^2> and the error on the <q> are calculated and
displayed at the end of the each run.

summery for data run 755,
total events : 122159 
rejected     : 37371 (due to the beam trip)

summery for data run 752,
total events : 202969 
rejected     : 2282

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@404 b9700a98-9937-0410-b777-f318d6eb5bff