The SIS3320 now has output tree support for accumulators (but no histos yet). For the sampling mode this is not meaningful at this point (only MPS level quantities can be stored).  I think that this is the same problem as in the tracking systems: how do you write variable length info to a tree.
A more general issue is at what level these triggered events should be saved.
Either a subbranch in the MPS branch could be used to store all triggered
events in that MPS period; or a separate tree could be used to store all
triggered events in a flat list, and store the associated MPS number.
Depending on whether we need any MPS info in the tracking mode (dependence
of the Q^2 on the helicity???) we should pick one.  Feedback anyone?

Trivial implementation of the new pure virtual methods for event cuts.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@406 b9700a98-9937-0410-b777-f318d6eb5bff