Together with Jeong Han we worked some more on the ROOT layer for the QwHitContainer.  It works now reasonably well, and is based on Jeong Han's TychoHit class for the VDCs at W&M.
[Writing]
QwHitContainerWrite (no arguments) will write a ROOT file (hitlist.root) which contains simulated
hit patterns.  A reasonable level of tree structure is maintained: every event contains a variable-
length list of hits, so no serialization needs to be done.

It should be easy to implement something like this for a generic CODA stream.

[Reading]
Don't complain about a lack of options here:
- QwHitContainerRead (no arguments) reads in the hitlist.root file using the QwAnalysis framework,
  for example to do tracking over previously saved hits.  It still does nothing more than dumping
  the hits to the screen.
- The macro Extensions/Macros/Tracking/ReadHitList.C loops over the hits in the ROOT file, and
  dumps them to the screen.  You will first need to load the QwHit shared libraries, for which
  instructions are on the first lines of LoadQwHit.C.  This will allow you to use all functions
  defined on QwHits and QwHitContainers in your interactive ROOT environment (e.g. select sublists
  of hits only in specific regions).
- You can open the hitlist.root file directly in ROOT and double-click in the TBrowser (i.e.
  plotting the fQwHits.fDriftDistance shows you a histogram of the drift distance for all hits in
  all events and you can see that there are two distinct contributions for region 2 and region 3).
- Finally, if you just load the hitlist.root file in ROOT as a command line argument (or in the
  file browser), you can now do simple(?) things like
    tree->Draw("hits.fQwHits.fDistance:hits.fQwHits.fElement","hits.fQwHits.fDirection==3")
  to plot the drift distance versus the wire number for U planes (and if you run LoadQwHit.C first,
  you can even use kDirectionU instead of the magic number 3).

We do not make sure yet that the correct pDetectorInfo pointer is set on read-back.  If we ever get
some fancy geometry lookup functions like we have on QwHitContainers, then we can set the pointer
correctly.

[Internals]
In QwHitRootContainer we use a TClonesArray of pointers to QwHits.  The big difference with the
regular QwHitContainer is that ROOT has polymorphic containers.  Basically, they just store a
pointer to whatever it happens to be pointing at.  This gives you a lot of freedom (... to mess
up very badly).  Since ROOT's file structure when saving containers is based around this poly-
morphism, it's difficult to save std::vector<QwHit> objects.  So we are still stuck using an
extra container class to save to ROOT.  If we really wanted to, we could transition QwHitContainer
to a version that uses ROOT containers (by mirroring the STL methods to their ROOT equivalents),
but we think that our frequent use of std::vector::iterators might cause problems (TIterator is
quite different, and might not allow for an in-place replacement, since it doesn't support increment
operators).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@438 b9700a98-9937-0410-b777-f318d6eb5bff