Definition of a helicity enumerator and map to signed int:   enum EQwHelicity = { kHelicityUndefined, kHelicityPositive, kHelicityNegative } and the map<EQwHelicity,int> kMapHelicity such that   kMapHelicity[kHelicityPositive] = +1,   kMapHelicity[kHelicityNegative] = -1,   kMapHelicity[kHelicityUndefined] = 0.
Right now the helicity in QwHelicity is defined as an int with values 0 and 1 for
negative and positive helicity, and -9999 for undefined helicity.  Helicity was
only defined internally to QwHelicity, so it couldn't be used outside of that class
unless by using the explicit numerical values.  I haven't changed anthing in the
QwHelicity, and at the moment this enum is at odds with what is used there!  But
maybe this enum could be used as an interface to the outside world.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@462 b9700a98-9937-0410-b777-f318d6eb5bff