Another attempt at making a 'transparent', browsable TTree with tracking event data...  It's not plugged into any release-quality executable yet but it sort-of works here on my laptop.  A cleaned-up port of qwsimtracking with demo should follow soon, time permitting.
The structure is as follows:

TFile file (could have a separate tree for slow or raw hits data)
|
TTree tree
|
|__ TBranch events of QwEvent (may be split, a.k.a. row-wise or column-wise)
    |
    |__ TClonesArray of QwHits (all QwHits have the same size)
    |  [so far identical to what is in QwHitRootContainer now]
    |
    |__ QwEventHeader with event number, time, etc...
    |
    |__ TClonesArray of QwPartialTrack (variable size objects...)
        |
        |__ TClonesArray of QwTreeLine (variable size objects...)
            | '
            | '
            |_'_ std::vector<QwHit> (I have no clue why TClonesArray
              '                                 doesn't work here)
              '
              '
This is basically what you see in the ROOT TBrowser, with some other
leafs floating around that I am evaluating for performance.  To some
extent you can do the double-clicking (yay!), BUT there is apparently
an inherent limitation in TTreeFormula that prevents this clicking
from working more than 2 levels deep.  The deepest QwHit list (in a
tree line) can not be easily evaluated.  And the plan was in fact to
go a level deeper by putting a layer TClonesArray of QwTrack between
the QwEvent level and the QwPartialTrack level, so we have to come up
with a solution or just dump the deep double-clicking.  (Limit of
TTreeFormula is indicated with dashed line.)

In concreto, this is what you cannot do:
$ tree->Draw("events.fQwPartialTracks.fQwTreeLines.fQwHits.fResidual")

That said, macros that load the shared object libraries of the class
definitions should just work, and compiled code should also work.

Some info about the TClonesArray implementation.  It actually keeps
a static TClonesArray in memory and the TClonesArray that gets written
is a subset of this static TClonesArray.  This has as big advantage
that you don't have to new/delete any objects (expensive), but it
requires that your objects have the same size so you can jump to a
particular one with a simple sizeof().  QwHits are simple enough that
they are always the same size.  Anything else has variable numbers of
'things' inside which makes them not as efficient for TClonesArrays.
But (nowadays) TclonesArray knows how to handle variable size objects.

Other options are:
- TObjArray: does not support browsing because it contains anything,
  even apples and oranges in the same array (well, pointers to any
  object that inherits from a TObject, to be precise).  Expensive in
  terms of new/delete.
- TRefArray: stores references only.  This could be useful if we would
  want to store references to the main (raw) hitlist in, for example,
  the tree lines.  However, we will likely write to the hits, and
  it is not nice to change the raw hits during tracking (especially
  because we might want to evaluate two settings simultaneously, and
  that would become difficult).

Probably related to the limitation in TTreeFormula, the browser also
doesn't seem to understand TClonesArrays that are deeper than two
levels.  However, std::vector works there (and might work everywhere
else too, what was the problem again with writing QwHitContainers
directly?).


I had to add a LinkDef for dictionary generation.  They should go in the
include directory, not in the dictionary directory.  You will need to do
a make config to get your dependencies right.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@463 b9700a98-9937-0410-b777-f318d6eb5bff