[ This is a private branch, just ignore if you don't use branches. ]
Modifications to create a diamond inheritance structure with a virtual
base class for the QwScanner (and other subsystems that will need to
be able to handle both parity and tracking data).

There are no modifications to the actual QwScanner code, only changes in
the inheritance structure.  I could as well have picked the lumi class.


                VQwSubsystem
                     /\
                    /  \
                   /    \
                  /      \
VQwSubsystemParity        VQwSubsystemTracking
         /        \      /        \
        /          \    /          \
       /            \  /            \
      /              \/              \
QwBeamLine       QwScanner       QwGasElectronMultiplier
  ...               ...            ...


To avoid ambiguity the inheritance in VQwSubsystemParity and
VQwSubsystemTracking is virtual, but that means that the
VQwSubsystem has to be explicitly initialized in the c'tors.

There are issues with the libraries (necessarily, considering
our build layout).  The new directory Common has no access to the
shared object libraries of the Tracking and Parity directories.
Add it by hand when linking: -lQwTracking -l QwParity

QwScannerTest is example code.  It creates a Parity array and a
Tracking array, both with the same QwScanner in it.  It crashes
when the arrays go out of scope, but that's normal when abusing
shared pointers.

Notes:
- dynamic_cast<T*>(ptr)  is preferable over  (T*) ptr
  otherwise upcasts won't work, you need to check for null
- virtual destructors are needed when a class will need to be
  extended (inherited)



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc@480 b9700a98-9937-0410-b777-f318d6eb5bff