Modified the evio "codaOpen" routine to allow us to open standard input and input pipes, in addition to regular files; this also allows us to stream compressed CODA data files through gunzip inline.  These routines are copied from a Hall C version of evio which was obtained from Steve Wood originally.
Added a set of cascaded file path checks to the QwEventBuffer::OpenDataFile,
such that when we try to open "filename" for read access, we searched for
it in this order:
     "filename"
     "DataDirectory"/"filename"
     "filename".gz
     "DataDirectory"/"filename".gz
The first instance which is found is the one which will be used; the
path used will be reported.
If we open "filename" for write access, the file will be opened without
modification of the path.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@501 b9700a98-9937-0410-b777-f318d6eb5bff