This implements a general logging facility for QwAnalysis(*).  The approach is to use an function call std::ostream QwLog(loglevel) everytime you want to write something:
  QwLog(2) << "this" << " that" << std::endl; // not recommended syntax

To make this more convenient there are a couple of macros defined in the
header: QwError, QwWarning, QwMessage (main messages), QwVerbose (more
messages), QwDebug.  And to get things working nicely we also need to
use QwLog::endl at the end:

  QwWarning << "Uh-oh" << QwLog::endl; // recommended syntax

  -> on the screen we get
    Warning: Uh-oh
  -> in the log file we get
    2009-11-26, 00:48:28 WW - Uh-oh

One global instance gQwLog can be used to set the log filename, the log
level for file output, and the log level for screen output:

  // Message logging facilities
  gQwLog.InitLogfile("QwTracking.log");
  gQwLog.SetScreenThreshold(QwLog::kMessage);
  gQwLog.SetFileThreshold(QwLog::kDebug);

This logging facility should pass things like std::hex as expected, but
I haven't tested this.  Also, it uses std::cerr but this is not necessary
since it knows about QwLog::flush.

For this version of the code the HDC class has been modified to use the
QwMessage stream.  All this took was a s/std::cout/QwMessage/ and a
s/std::endl/QwLog::endl/.


There are a few things that can be improved.
- QwMessage << "this" << QwLog::endl << "that" << QwLog::endl; does not
  reprint the time/date lines in the file.
- Printing time/date is maybe not that useful, but we could modify this
  to print run/event.  For this to work gQwLog would need a hook to the
  event buffer.
- Inside objects it would be possible to print fSystemName one each line.
  This would however require fSystemName to be present, which is not the
  case in main().
- More fine-grained logging control would be useful, for example treating
  all BCM objects with log level kDebug while keeping all other systems
  at log level kWarning.  One approach I am considering is to add a debug
  level escalation variable to the VQwSubsystem class.  When this is set
  to 1 all messages get escalated 1 log level for that object only.  Here
  again the problem is that this should also work outside of objects
  derived from VQwSubsystem.


(*) It is a shameless rip-off of some Hermes code.  But even there it
probably wasn't completely original...



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@530 b9700a98-9937-0410-b777-f318d6eb5bff