After the logger facility yesterday, my Thanksgiving (late) evening was spent writing a generic options class.  As suggested by Paul it uses boost::program_ options (or namespace alias po for short).  One global object is responsible for the argc/argv stuff and a main options file, but one could theoretically just instantiate extra QwOptions objects as needed on a class by class basis (preferably static to avoid overhead), and let them parse only config files.
The global object is necessarily instantiated in main(argc, argv) as

  QwOptions gQwOptions(argc, argv, "qweak.config");

with an (at the moment unused) options file as third argument.  You can then
access the options using a key/value system.  Using the example of yesterday's
logging facility:

  gQwLog.InitLogfile(gQwOptions.GetStringValue("logfile"));
  gQwLog.SetScreenThreshold(QwLog::QwLogLevel(gQwOptions.GetIntValue("loglevel-screen")));
  gQwLog.SetFileThreshold(QwLog::QwLogLevel(gQwOptions.GetIntValue("loglevel-file")));

[There's an interesting chicken and egg problem here: you need a logger to
notify problems in option parsing, but you need to parse the options before
you know what and where to log...]

This now would allow us to run the analyzer as follows

  bin/qwtracking --logfile test.log --loglevel-screen 0 --loglevel-file 3 --run 100:105

to get only errors on the screen, but everything up to verbose messages in the
log file, and to process runs 100 until 105.

Ranges are implemented as in QwCommandLine, and you can access them using

  std::pair<int,int> GetIntValuePair(string key);
  int GetIntValuePairFirst(string key);
  int GetIntValuePairLast(string key);

In the run loop this could be used as

  for (UInt_t run =  (UInt_t) gQwOptions.GetIntValuePair("run").first;
              run <= (UInt_t) gQwOptions.GetIntValuePair("run").second; run++) {
or
  for (UInt_t run =  (UInt_t) gQwOptions.GetIntValuePairFirst("run");
              run <= (UInt_t) gQwOptions.GetIntValuePairLast("run"); run++) {


Further functions provided:
- test whether a flag is set, or whether a numeric option is present:
  bool GetBoolValue(string key);
- get a double value:
  double GetDoubleValue(string key);

At the moment all available options need to be set in the QwOptions.cc file.
The constructor is where the parsing is done, so it needs to know all options
in advance, but I would assume that we can split that out and allow for classes
to add keywords and re-parse argv and the config file.  Those keywords would
not show up in the usage instructions, since we're already past that then.
If you have read this far and you see a need for this, let me know and I will
think about it some more.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@532 b9700a98-9937-0410-b777-f318d6eb5bff