Here are some more modifications to the QwOptions class.  It is now relatively easy to add options to a user class.  It can now also use configuration files for options.
In the main() routine, you could put 'somewhere' the following lines (sorry for
the API change, also no need anymore to define the gQwOptions object):

  // Set the command line arguments and configuration file
  gQwOptions.SetCommandLine(argc, argv);
  gQwOptions.SetConfigFile("qwsimtracking.conf");
  // Define some options
  gQwOptions.AddOptions()("run,r", po::value<string>(), "run range in format #[:#]");
  gQwOptions.AddOptions()("event,e", po::value<string>(), "event range in format #[:#]");

This will make a copy of the command line arguments (for use in other objects
that don't have access to the original argc and argv).  You can also add your
own options (note the first empty parentheses, they are important).  In this
case the options -r/--run and -e/--event are defined with string as type
(because they have that funny range format).

And then you continue almost as before with:

  gQwOptions.GetValue<string>("logfile")

but this is now a template function for generality (and for adventure).

The good part is that you can use the AddOptions() function everywhere you
include QwOptions.h (preferably in the constructors).  This will then provide
the options to everything that comes later.  They will probably not show up
in the --help text, because that gets generated on the first GetValue<>() call
and that is probably before your constructors get called.  No-one has invented
future-aware programming yet.

Parsing of options happens when an option is requested via GetValue<>(), but
only if something has changed (new argv added, new config file, new options
defined).

If you want to use these options in a class, prefix your options with the
class name (or some shorter mnemonic).  For example: HDC.foo.  This will then
be in a separate section [HDC] of the config file.

To test this, run bin/qwsimtracking without options:
  Welcome to the Qweak analyzer code.

  Allowed options:
    --help                     print this help message
    --usage                    print this help message
    --logfile arg              log file
    --loglevel-file arg (=4)   log level for file output
    --loglevel-screen arg (=2) log level for screen output
    -r [ --run ] arg           run range in format #[:#]
    -e [ --event ] arg         event range in format #[:#]
then create a text file qwsimtracking.conf with content
  logfile = qwsimtracking.log
  loglevel-file = 5
  loglevel-screen = 2
  run = 100
and run 'bin/qwsimtracking -e 0:10' to run over 11 events and to write to that
logfile with specified log levels.  This does not use QwCommandLine anymore.
The logging options are defined in the different class QwLog.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@534 b9700a98-9937-0410-b777-f318d6eb5bff